feat: Add intersection observer script for TOC highlighting
This looks for the sections created by rehype-sectionize. Uses the
intersection observer web API to attach observers to each section in the
mdx content. When the content is in the viewport it will get the <a> tag
element in the TOC with the same HREF and add highlighting styles to the
class. This script is based off a blog post by Reza Zahedi.
https://rezahedi.dev/blog/create-table-of-contents-in-astro-and-sectionize-the-markdown-content