Fix: Fix locale change on homepage caused by build format
Astro-pagefind package requires you to change the build format from the
default directory to "file". This caused massive issues in the build,
and also changing the locale on the homepage when in preview. So instead
of en.html and ja.html it outputs a en/index.html and ja/index.html in
the build process